refactor: move automation handlers into chats.go, fix review issues
Move all chat automation HTTP handlers from the standalone
chatautomations.go into chats.go to keep the chat API surface in
one file, matching the project convention.

Fixes from code review:
- Add dbauthz.AsSystemRestricted(ctx) for the unauthenticated
  webhook endpoint's DB call.
- Add cron schedule validation via cron.Weekly() in create handler.
- Fix swagger: update [patch] → [put], 201 → 202 on webhook.
- Add EXPERIMENTAL comments to all automation handlers.
- Use dbtime.Now() instead of time.Now() in executor.
- Remove dead ChatCreator interface from executor.go.
- Log errors instead of silently discarding in executor failure path.
- Name executor logger ("autochat").
- Validate non-empty webhook secret in VerifyWebhookSignature.
- Normalize hex case in signature verification.
- Convert promptFuncMap from mutable package-level var to function.
- Use literal expected value in FuncPipeline test.
- Add empty-string test case for MaskSecret.
- Fix collapsed lines and indentation in coderd.go (gofmt).
